Add smiley support (Fix #1)
frowny.el now supports smileys (i.e., ":)") as well as frownies.  Why anyone
would want to express happiness is beyond this author, but frowny.el is for the
people.

In order to do so, `frowny-self-insert' is now deprecated, and will be removed
in a future version.  Users are encouraged to convert their code to use
`frowny-self-insert-frowny' instead.

In addition, the key definition to insert a smiley has been added to
`frowny-mode-map'.